requirements.txt
annotated-types==0.7.0
argostranslate==1.10.0
asgiref==3.10.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
ctranslate2==4.6.0
cymem==2.0.11
Django==5.2.7
djangorestframework==3.16.1
emoji==2.15.0
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==2.3.4
packaging==25.0
preshed==3.0.10
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rich==14.2.0
sacremoses==0.1.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.4.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stanza==1.10.1
sympy==1.14.0
thinc==8.3.6
torch==2.9.0
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
xx_sent_ud_sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.8.0/xx_sent_ud_sm-3.8.0-py3-none-any.whl#sha256=1b5fb9321723b395f76458d2a13b7a3673c537eb58c0a9df3778f4e0950180b4
